# Omit executing methods/blocks that are side-effect free