--- a/tool/ruby_vm/views/insns_info.inc.erb
+++ b/tool/ruby_vm/views/insns_info.inc.erb
@@ -15,5 +15,7 @@
 <%= render 'insn_name_info' %>
 <%= render 'insn_len_info' %>
 <%= render 'insn_operand_info' %>
+<%= render 'insn_purity_helpers' %>
 <%= render 'attributes' %>
 <%= render 'insn_stack_increase' %>
+<%= render 'insn_purity' %>