--- a/vm.c
+++ b/vm.c
@@ -324,6 +324,7 @@ rb_serial_t
 rb_next_class_serial(void)
 {
     rb_serial_t class_serial = NEXT_CLASS_SERIAL();
+    INC_GLOBAL_TIMESTAMP();
     mjit_add_class_serial(class_serial);
     return class_serial;
 }
@@ -338,6 +339,7 @@ rb_vm_t *ruby_current_vm_ptr = NULL;
 rb_execution_context_t *ruby_current_execution_context_ptr = NULL;
 rb_event_flag_t ruby_vm_event_flags;
 rb_event_flag_t ruby_vm_event_enabled_flags;
+rb_serial_t ruby_vm_global_timestamp = 1;
 rb_serial_t ruby_vm_global_method_state = 1;
 rb_serial_t ruby_vm_global_constant_state = 1;
 rb_serial_t ruby_vm_class_serial = 1;
@@ -414,6 +416,7 @@ static VALUE
 vm_stat(int argc, VALUE *argv, VALUE self)
 {
     static VALUE sym_global_method_state, sym_global_constant_state, sym_class_serial;
+    static VALUE sym_global_timestamp;
     VALUE arg = Qnil;
     VALUE hash = Qnil, key = Qnil;
 
@@ -431,6 +434,7 @@ vm_stat(int argc, VALUE *argv, VALUE self)
 
     if (sym_global_method_state == 0) {
 #define S(s) sym_##s = ID2SYM(rb_intern_const(#s))
+	S(global_timestamp);
 	S(global_method_state);
 	S(global_constant_state);
 	S(class_serial);
@@ -443,6 +447,7 @@ vm_stat(int argc, VALUE *argv, VALUE self)
     else if (hash != Qnil) \
 	rb_hash_aset(hash, sym_##name, SERIALT2NUM(attr));
 
+    SET(global_timestamp, ruby_vm_global_timestamp);
     SET(global_method_state, ruby_vm_global_method_state);
     SET(global_constant_state, ruby_vm_global_constant_state);
     SET(class_serial, ruby_vm_class_serial);