--- a/vm_core.h
+++ b/vm_core.h
@@ -275,6 +275,9 @@ struct rb_call_cache {
 	enum method_missing_reason method_missing_reason; /* used by method_missing */
 	int inc_sp; /* used by cfunc */
     } aux;
+
+    rb_serial_t updated_at;
+    VALUE purity;
 };
 
 #if 1
@@ -455,6 +458,10 @@ struct rb_iseq_constant_body {
     long unsigned total_calls; /* number of total calls with `mjit_exec()` */
     struct rb_mjit_unit *jit_unit;
     char catch_except_p; /* If a frame of this ISeq may catch exception, set TRUE */
+
+    rb_serial_t updated_at;
+    VALUE purity;
+    bool iter_lev;
 };
 
 /* T_IMEMO/iseq */