--- a/vm_insnhelper.c
+++ b/vm_insnhelper.c
@@ -11,6 +11,7 @@
 /* finish iseq array */
 #include "insns.inc"
 #ifndef MJIT_HEADER
+#define INSIDE_VM_INSNHELPER_C
 #include "insns_info.inc"
 #endif
 #include <math.h>
@@ -1311,6 +1312,7 @@ rb_vm_search_method_slowpath(const struct rb_call_info *ci, struct rb_call_cache
     cc->method_state = GET_GLOBAL_METHOD_STATE();
     cc->class_serial = RCLASS_SERIAL(klass);
 #endif
+    CC_RESET_PURITY(cc);
 }
 
 static void
@@ -2059,6 +2061,7 @@ vm_call_opt_send(rb_execution_context_t *ec, rb_control_frame_t *reg_cfp, struct
 
     cc->me = rb_callable_method_entry_with_refinements(CLASS_OF(calling->recv), ci->mid, NULL);
     ci->flag = VM_CALL_FCALL | VM_CALL_OPT_SEND;
+    CC_RESET_PURITY(cc);
     return vm_call_method(ec, reg_cfp, calling, ci, cc);
 }
 
@@ -2154,6 +2157,7 @@ vm_call_zsuper(rb_execution_context_t *ec, rb_control_frame_t *cfp, struct rb_ca
 	cc->me->def->body.refined.orig_me) {
 	cc->me = refined_method_callable_without_refinement(cc->me);
     }
+    CC_RESET_PURITY(cc);
     return vm_call_method_each_type(ec, cfp, calling, ci, cc);
 }
 
@@ -2295,6 +2299,7 @@ vm_call_method_each_type(rb_execution_context_t *ec, rb_control_frame_t *cfp, st
       case VM_METHOD_TYPE_ALIAS:
 	cc->me = aliased_callable_method_entry(cc->me);
 	VM_ASSERT(cc->me != NULL);
+	CC_RESET_PURITY(cc);
 	return vm_call_method_each_type(ec, cfp, calling, ci, cc);
 
       case VM_METHOD_TYPE_OPTIMIZED:
@@ -2341,12 +2346,14 @@ vm_call_method_each_type(rb_execution_context_t *ec, rb_control_frame_t *cfp, st
 		}
 	    }
 	    cc->me = ref_me;
+	    CC_RESET_PURITY(cc);
 	    if (ref_me->def->type != VM_METHOD_TYPE_REFINED) {
 		return vm_call_method(ec, cfp, calling, ci, cc);
 	    }
 	}
 	else {
 	    cc->me = NULL;
+	    CC_RESET_PURITY(cc);
 	    return vm_call_method_nome(ec, cfp, calling, ci, cc);
 	}
 
@@ -2358,6 +2365,7 @@ vm_call_method_each_type(rb_execution_context_t *ec, rb_control_frame_t *cfp, st
 	    VALUE klass = RCLASS_SUPER(cc->me->defined_class);
 	    cc->me = klass ? rb_callable_method_entry(klass, ci->mid) : NULL;
 	}
+	CC_RESET_PURITY(cc);
 	return vm_call_method(ec, cfp, calling, ci, cc);
       }
     }
@@ -3820,6 +3828,39 @@ vm_opt_regexpmatch2(VALUE recv, VALUE obj)
     else {
 	return Qundef;
     }
+}
+
+/* @shyouhei wonders: where is the right place to define this function? */
+bool
+vm_whether_we_can_skip_this_call_handler_p(vm_call_handler h)
+{
+    /* We cannot write this function using a switch() because a case
+     * label cannot be a function pointer. */
+
+    /* Because there are lots of call handlers such as
+     * vm_call_iseq_setup_normal_0start_2params_4locals(),
+     * we cannot whitelist them all. See also
+     * tools/mk_call_iseq_optimzied.rb */
+    static const vm_call_handler blacklist[] = {
+        NULL,
+        vm_call_attrset,
+        vm_call_cfunc,
+        vm_call_general,
+        vm_call_ivar,
+        vm_call_method_missing,
+        vm_call_opt_block_call,
+        vm_call_opt_call,
+        vm_call_opt_send,
+        vm_call_super_method,
+    };
+    int i;
+
+    for (i = 0; i < numberof(blacklist); i++) {
+        if (h == blacklist[i]) {
+            return false;
+        }
+    }
+    return true;
 }
 
 rb_event_flag_t rb_iseq_event_flags(const rb_iseq_t *iseq, size_t pos);