… insns.def
… tool/ruby_vm/views/_insn_purity_helpers.erb