@k0kubun 2018-09-04
> purity: describes if the instruction has a side effect or not.

Could you elaborate the definition of "side effect" more? At first I imagined all instructions that alter *any* state are not pure, but it seems that you consider state changes *inside the method* as pure, like `setlocal` for a local variable in the method and almost all instructions that modify stack pointer.

So, to be more precise, does "purity" mean that the instruction does not alter states outside the method which has the instruction?

<hr>

During an insn is pure, can we think arbitrary method call does not happen? If so, we may be able to skip program counter motion for pure insns on MJIT.

<hr>

Also, how did you detect / implement the purity flags? Reading all related code? After merging this, every time we modify any C function, how can we know all instructions that may call the modified function are pure?

If there were assertions enabled by `VM_CHECK_MODE` that ensure side effect is not happening inside pure insns, it would be very helpful to maintain the flags.