@shyouhei 2018-09-05
My initial design about purity was that if an instruction manipulates anything other than the stack pointer and values on the stack, that is not pure.  Later, this turned out to be insufficient.  For instance let's see this example:

```ruby
def foo
  yield if defined? yield
end

def bar
  x = 1
  foo do
    x = 2
  end
  puts x
end

bar
```

Here, `bar` is not pure (obvious; it does IO).  OTOH at a glance `foo` seems to be pure because there is no "side effect"-ish instructions inside.  But if you skip the call to the method, `bar` behaves differently.  So _this_ call to `foo` is not pure.  What makes things complicated is that `foo` _could_ be called in a pure way e.g. when called without a block.  The `foo` call in the above example is not pure because its yielding block is not pure.

So the situation is complicated.  Currently, I define pure operations as the following:

- Pushing new values onto the stack, and eventually popping them, are pure operations.
- Reading the values that exist in the stack of any depth is a pure operation.  Writing is not; local variable assignment is pure if and only if its "level" is zero.
- A method call is pure if we can say all the instructions it and its block contains are recursively pure.
- Anything other than these are not pure.

> does "purity" mean that the instruction does not alter states outside the method which has the instruction?

This is basically a right observation except that the definition of "states" is not that obvious.  My one above makes it clear that we are focusing on the stack.

> During an insn is pure, can we think arbitrary method call does not happen?

This is in general NG because `invokesuper` etc. can be pure.  But for the majority of "normal" instructions such as `push` / `pop`, this is true.

You can say that an instruction has no method call if it (1) is pure and (2) has no CALL_CACHE operands.  Of course, there are instructions without method calls which do not meet these criteria.  But if one satisfies both of them, there are no method calls inside.

> Also, how did you detect / implement the purity flags? Reading all related code?

Yes.

> After merging this, every time we modify any C function, how can we know all instructions that may call the modified function are pure?

I agree this is a problem.  However, I wonder how often there will be situations like a pure instruction to suddenly become not pure.  Purity is a "nature" of an instruction that is (I guess) consistent over time.

I don't think it's easy to _ensure_ detection of every and all of purity breakage (sounds like a kind of proof-related activity).  But I agree there can be better-than-nothing checks, like detection of method calls inside of an instruction.  I'll try.