@k0kubun 2018-09-05
> Pushing new values onto the stack, and eventually popping them, are pure operations.
> Reading the values that exist in the stack of any depth is a pure operation. Writing is not; local variable assignment is pure if and only if its "level" is zero.
> A method call is pure if we can say all the instructions it and its block contains are recursively pure.
> Anything other than these are not pure.

That made my understanding of your intention much clearer. Thank you.

> You can say that an instruction has no method call if it (1) is pure and (2) has no CALL_CACHE operands. Of course, there are instructions without method calls which do not meet these criteria. But if one satisfies both of them, there are no method calls inside.

Okay, I understand there's purity that may be changed by method cache. If we use purity for optimization, we need to invalidate the optimization outcome when it's made not pure.

> I wonder how often there will be situations like a pure instruction to suddenly become not pure. Purity is a "nature" of an instruction that is (I guess) consistent over time.

My main concern is that I'm not sure whether any `rb_*` C function calls `rb_funcall` or not and we would need to think about it every time we use `rb_*` C functions in vm_insnhelper.c. Since my concern is mainly for underlying rb_funcall, "detection of method calls inside of an instruction" would solve most of it.