@shyouhei 2018-09-07
#1947 is the thing that check "can we think arbitrary method call does not happen" problem.