@ko1 2018-09-09
(1) Discussion of drawbacks

Could you give us the drawbacks of your suggestion?
I haven't check your source code in details, but it seems to update and to check something.
If your optimization is limited, I think it shouldn't be introduced.
Some benchmarks shows speed-down.

I agree that this kind of analysis is needed for future optimization.
But I don't think we need to introduce such overheads if it doesn't show good (or bad) performance. 

(2) Real benchmarks

Could you show us the results with real benchmarks?
For example, rdoc or discourse benchmark <https://github.com/discourse/discourse/blob/master/script/bench.rb> can help you.

(3) Estimation of real benchmarks

Could you count the "pure" iseqs on real applications?
At least "count" can be two types: static counts (count at compile time) and dynamic counts (count how many pure iseqs are invoked). Other "counts" can be possible.
And you can show some estimation how your proposal is effective on the current optimization and future optimizations. This kind of estimation will increase your idea's persuasion.

(comments) How to maintain this kind of attribute when we modify insns.def? Any tests are there? Sorry If I missed on your explanation.

Thanks,
Kochi